Devided in two classes
